C - Functions, nested loops: task 14. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A. fixing overflow errors

+++ functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	int ii, old, new, newnew;
+	long int ii, old, new, newnew;
 
 	old = 1;
 	new = 1;
